多体larmor運動の可視化
